requirements.txt
streamlit>=1.34
transformers>=4.42
accelerate>=0.33
torch>=2.2
pydantic>=2.7
python-dateutil>=2.9
einops>=0.7
safetensors>=0.4